Remove special handling for compiler flag -Wreserved-identifier + most exclusions
Compiler flag -Wreserved-identifier checks if identifiers use a reserved
name, e.g. start with an underscore. It is available in Clang >= 12 but
not in GCC. Because of this and because not all versions of apple-clang
support it, it was only conditionally activated using
check_cxx_compiler_flag / HAS_RESERVED_IDENTIFIER.

But:
1. -Wreserved-identifier is enabled already by -Weverything when using
   Clang >=12 (= the minimum version in ClickHouse). Also, -Weverything
   is the default in ClickHouse builds (proof: cmake/warnings.cmake). So
   everyone using Clang >= 12 will be warned.
2. For everyone else (gcc/appleclang): clang-tidy finds the same kind of
   issues using check "bugprone-reserved-identifier" which is also
   activated by default.

This allows to drop some special handling and developers will still get
the warning, at the lastest when pushing to GitHub. I also trimmed down
some unnecessary exclusions for -Wreserved-identifier.